Add modules in module folder
