+++ modules/library.js
@@ -1,0 +1,20 @@
+export default class Library {
+  constructor() {
+    this.books = [];
+  }
+
+  addBook(book) {
+    this.books.push(book);
+  }
+
+  removeBook(book) {
+    const index = this.books.indexOf(book);
+    if (index !== -1) {
+      this.books.splice(index, 1);
+    }
+  }
+
+  getBooks() {
+    return this.books;
+  }
+}
